Completed Nushell setup script
A gorgeous 50 lines of scripting to replace the hundreds I had before

+++ config-setup.nu
@@ -1,25 +1,50 @@
 #/usr/bin/nu
 
+# `dir`: name of directory to symlink to `~/.config/`
+# `deps` (optional): system programs that the config depends on
+let configs = [
+  { dir: "nvim", deps: ["nvim", "rg"]},
+  # { dir: "waybar", deps: ["waybar", "pavucontrol", "python3"] },
+  { dir: "fuzzel", deps: ["fuzzel"] },
+  # { dir: "hypr", deps: ["hyprland"] },
+  { dir: "zed", deps: ["zeditor"] },
+  # { dir: "ghostty", deps: ["ghostty"] }, # Also needs `ttf-jetbrains-mono-nerd` package
+  { dir: "kitty", deps: ["kitty", "nu"] }, # Also needs `ttf-jetbrains-mono-nerd` package
+  { dir: "nushell", deps: ["nu", "pokeget"] },
+  { dir: "niri", deps: ["niri", "ironbar", "thunar", "fuzzel", "swaync", "swaybg", "xwayland-satellite", "wpctl", "brightnessctl"] },
+  { dir: "ironbar", deps: ["ironbar", "playerctl"] },
+]
+
 let repo_dir = ($env.PWD)
 let conf_home = ($env.HOME | path join ".config")
 
-# See `configs.nuon` to customize setup
-let configs = open configs.nuon
+let missing_progs = (
+  $configs
+  | each {|conf|
+    print $"Setting up (ansi blue)($conf.dir)(ansi reset)..."
 
-for conf in $configs {
-  let pkg_name = ($conf.package? | default $conf.program)
+    # Find missing commands
+    let $missing = ($conf.deps? | default []) | where {|p| (which $p | is-empty) }
+    for prog in $missing {
+      print $"    (ansi yellow_bold)Warning:(ansi reset) Command (ansi blue)($prog)(ansi reset) not found on this system."
+    }
 
-  if (which $conf.program | is-empty) {
-    print (ansi yellow_bold) + "WARNING:" + (ansi reset) + $" ($conf.program) not found. Try installing ($conf.package)"
+    # Symlinking operations
+    let source = ($repo_dir | path join ".config" | path join $conf.dir)
+    let target = ($conf_home | path join $conf.dir)
+
+    if ($target | path exists) { rm -rf $target }
+    ln -s $source $target
+
+    # Return missing commands
+    $missing
   }
+  | flatten
+  | uniq
+)
 
-  let source = ($repo_dir | path join ".config" | path join $conf.dir)
-  let target = ($conf_home | path join $conf.dir)
-
-  if ($target | path exists) {
-    rm -rf $target
-  }
-
-  ln -s $source $target
+print $"(ansi green)Done(ansi reset)." 
+if (not ($missing_progs | is-empty)) {
+  print $"Missing programs: ($missing_progs | str join ' ')"
 }
